corretto bug

+++ Script Sql/script_libreria.sql
@@ -7,5 +7,5 @@ titolo varchar(50) not null,
 autore varchar(50) not null,
 genere varchar(50) not null,
 data_pubblicazione date not null,
-prezzo decimal
+prezzo double
 );
